core: simplify ontology materializer implementation
Decompose the in-memory materialization provider and root storage locking into focused modules. Consolidate projection execution, identity, watermark, ordering, and correlation helpers while removing redundant session and edit workflow code. Projection telemetry now validates replay and commit in one fenced transaction.
