+++ packages/core/src/materializer/execution/run-correlation.ts
@@ -1,6 +1,51 @@
-import { MaterializationConflictError } from "../../materialization/errors"
+import {
+  MaterializationConflictError,
+  MaterializationValidationError,
+} from "../../materialization/errors"
 import type { ProjectionDefinition, ResolvedProjection } from "../../projections/types"
-import type { ProjectionMaterializationRunRecord } from "../../storage"
+import {
+  isProjectionMaterializationRunStorage,
+  type ProjectionMaterializationRunRecord,
+  type ProjectionMaterializationRunStorage,
+  type ProjectionRunMaterializationIdentity,
+  type Storage,
+} from "../../storage"
+
+export interface AssertProjectionExecutionInput {
+  readonly projectId: string
+  readonly projectionRunId: string
+  readonly executionToken: string
+  readonly identity: ProjectionRunMaterializationIdentity
+  readonly resolved: ResolvedProjection<ProjectionDefinition>
+  readonly capabilityErrorMessage?: string
+}
+
+export interface AssertedProjectionExecution {
+  readonly projectionRuns: ProjectionMaterializationRunStorage
+  readonly run: ProjectionMaterializationRunRecord
+}
+
+export async function assertProjectionMaterializationExecution(
+  storage: Pick<Storage, "projectionRuns">,
+  input: AssertProjectionExecutionInput
+): Promise<AssertedProjectionExecution> {
+  if (!isProjectionMaterializationRunStorage(storage.projectionRuns)) {
+    throw new MaterializationValidationError(
+      input.capabilityErrorMessage ??
+        "Storage transaction does not provide projection run capabilities."
+    )
+  }
+
+  const projectionRuns = storage.projectionRuns
+  const run = await projectionRuns.assertMaterializationExecution({
+    id: input.projectionRunId,
+    projectId: input.projectId,
+    executionToken: input.executionToken,
+    identity: input.identity,
+  })
+  assertProjectionRunTargets(run, input.resolved)
+  return { projectionRuns, run }
+}
 
 export function assertProjectionRunTargets(
   run: ProjectionMaterializationRunRecord,
