+++ packages/core/src/storage/ontology/in-memory/shared-state.ts
@@ -1,3 +1,4 @@
+import { MaterializationValidationError } from "../../../materialization/errors"
 import type { OntologyMaterializationOrigin } from "../../../materialization/model"
 import type { OntologyCommitOriginSelector, OntologyCommitRecord } from "../commits"
 import type {
@@ -130,4 +131,31 @@ export function restoreOntologyState(
     targetMap.clear()
     for (const [entryKey, value] of sourceMap) targetMap.set(entryKey, value)
   }
+}
+
+/** Insert `value` into a sorted, size-bounded accumulator, dropping the largest entry past `limit`. */
+export function insertBounded<T>(
+  values: T[],
+  value: T,
+  limit: number,
+  compare: (left: T, right: T) => number
+): void {
+  let index = values.findIndex((candidate) => compare(value, candidate) < 0)
+  if (index < 0) index = values.length
+  values.splice(index, 0, value)
+  if (values.length > limit) values.pop()
+}
+
+export function assertNonblank(value: string, label: string): void {
+  if (typeof value !== "string" || value.trim().length === 0) {
+    throw new MaterializationValidationError(`${label} must be nonblank.`)
+  }
+}
+
+export function assertTimestamp(value: string, label: string): number {
+  const milliseconds = Date.parse(value)
+  if (!Number.isFinite(milliseconds)) {
+    throw new MaterializationValidationError(`${label} must be a valid timestamp.`)
+  }
+  return milliseconds
 }
